+++ pymonetdb/__init__.py
@@ -19,32 +19,24 @@ from pymonetdb.sql.connections import Connection
 from pymonetdb.sql.pythonize import *
 from pymonetdb.exceptions import *
 
-
 try:
     __version__ = pkg_resources.require("pymonetdb")[0].version
 except pkg_resources.DistributionNotFound:
     __version__ = "1.0rc"
 
-
-__all__ = ["sql", "mapi", "exceptions"]
 apilevel = "2.0"
 threadsafety = 0
 paramstyle = "pyformat"
 
-__all__ = ['BINARY', 'Binary', 'connect', 'Connection', 'DATE',
-           'Date', 'Time', 'Timestamp', 'DateFromTicks', 'TimeFromTicks',
-           'TimestampFromTicks', 'DataError', 'DatabaseError', 'Error',
-           'FIELD_TYPE', 'IntegrityError', 'InterfaceError', 'InternalError',
-           'MySQLError', 'NULL', 'NUMBER', 'NotSupportedError', 'DBAPISet',
-           'OperationalError', 'ProgrammingError', 'ROWID', 'STRING', 'TIME',
-           'TIMESTAMP', 'Set', 'Warning', 'apilevel', 'connect',
-           'connections', 'constants', 'cursors', 'debug', 'escape',
-           'escape_dict', 'escape_sequence', 'escape_string',
-           'get_client_info', 'paramstyle', 'string_literal', 'threadsafety',
-           'version_info']
+__all__ = ['sql', 'mapi', 'exceptions', 'BINARY', 'Binary', 'connect', 'Connection', 'DATE', 'Date', 'Time',
+           'Timestamp', 'DateFromTicks', 'TimeFromTicks', 'TimestampFromTicks', 'DataError', 'DatabaseError', 'Error',
+           'IntegrityError', 'InterfaceError', 'InternalError', 'NUMBER', 'NotSupportedError', 'OperationalError',
+           'ProgrammingError', 'ROWID', 'STRING', 'TIME', 'Warning', 'apilevel', 'connect', 'paramstyle',
+           'threadsafety']
 
 
 def connect(*args, **kwargs):
     return Connection(*args, **kwargs)
+
 
 connect.__doc__ = Connection.__init__.__doc__
