+++ pymonetdb/sql/cursors.py
@@ -6,15 +6,13 @@
 
 import logging
 from collections import namedtuple
-import tempfile
-import re
 import pickle
-import pdb
-
+from typing import Optional, Dict
+from pymonetdb.sql.debug import debug
 from pymonetdb.sql import monetize, pythonize
 from pymonetdb.exceptions import ProgrammingError, InterfaceError
 from pymonetdb import mapi
-from six import u, PY2
+from six import PY2, ensure_binary
 
 logger = logging.getLogger("pymonetdb")
 
@@ -125,6 +123,7 @@ class Cursor(object):
         self.connection = None
 
     def execute(self, operation, parameters=None):
+        # type: (str, Optional[Dict]) -> int
         """Prepare and execute a database operation (query or
         command).  Parameters may be provided as mapping and
         will be bound to variables in the operation.
@@ -136,13 +135,8 @@ class Cursor(object):
         # clear message history
         self.messages = []
 
-        # convert to utf-8
         if PY2:
-            if type(operation) == unicode:
-                # don't decode if it is already unicode
-                operation = operation.encode('utf-8')
-            else:
-                operation = u(operation).encode('utf-8')
+            operation = ensure_binary(operation)
 
         # set the number of rows to fetch
         if self.arraysize != self.connection.replysize:
@@ -158,8 +152,7 @@ class Cursor(object):
         query = ""
         if parameters:
             if isinstance(parameters, dict):
-                query = operation % dict([(k, monetize.convert(v))
-                                          for (k, v) in parameters.items()])
+                query = operation % {k: monetize.convert(v) for (k, v) in parameters.items()}
             elif type(parameters) == list or type(parameters) == tuple:
                 query = operation % tuple(
                     [monetize.convert(item) for item in parameters])
@@ -191,81 +184,6 @@ class Cursor(object):
         self.rowcount = count
         return count
 
-    def __exportparameters(self, ftype, fname, query, quantity_parameters,
-                           sample):
-        """ Exports the input parameters of a given UDF execution
-            to the Python process. Used internally for .debug() and
-            .export() functions.
-        """
-
-        # create a dummy function that only exports its parameters
-        # using the pickle module
-        if ftype == 5:
-            # table producing function
-            return_type = "TABLE(s STRING)"
-        else:
-            return_type = "STRING"
-        if sample == -1:
-            export_function = """
-                CREATE OR REPLACE FUNCTION export_parameters(*)
-                RETURNS %s LANGUAGE PYTHON
-                {
-                    import inspect
-                    import pickle
-                    frame = inspect.currentframe();
-                    args, _, _, values = inspect.getargvalues(frame);
-                    dd = {x: values[x] for x in args};
-                    del dd['_conn']
-                    return pickle.dumps(dd);
-                };""" % return_type
-        else:
-            export_function = """
-                CREATE OR REPLACE FUNCTION export_parameters(*)
-                RETURNS %s LANGUAGE PYTHON
-                {
-                import inspect
-                import pickle
-                import numpy
-                frame = inspect.currentframe();
-                args, _, _, values = inspect.getargvalues(frame);
-                dd = {x: values[x] for x in args};
-                del dd['_conn']
-                result = dict()
-                argname = "arg1"
-                x = numpy.arange(len(dd[argname]))
-                x = numpy.random.choice(x,%s,replace=False)
-                for i in range(len(dd)-2):
-                    argname = "arg" + str(i + 1)
-                    result = dd[argname]
-                    aux = []
-                    for j in range(len(x)):
-                        aux.append(result[x[j]])
-                    dd[argname] = aux
-                    print(dd[argname])
-                print(x)
-                return pickle.dumps(dd);
-                };
-                """ % (return_type, str(sample))
-
-        if fname not in query:
-            raise Exception("Function %s not found in query!" % fname)
-
-        query = query.replace(fname, 'export_parameters')
-        query = query.replace(';', ' sample 1;')
-
-        self.execute(export_function)
-        self.execute(query)
-        input_data = self.fetchall()
-        self.execute('DROP FUNCTION export_parameters;')
-        if len(input_data) <= 0:
-            raise Exception("Could not load input data!")
-        arguments = pickle.loads(str(input_data[0][0]))
-
-        if len(arguments) != quantity_parameters + 2:
-            raise Exception("Incorrect amount of input arguments found!")
-
-        return arguments
-
     def export(self, query, fname, sample=-1, filespath='./'):
         """ Exports a Python UDF and its input parameters to a given
             file so it can be called locally in an IDE environment.
@@ -284,7 +202,7 @@ class Cursor(object):
             ORDER BY args.number;""" % fname)
         input_names = self.fetchall()
         quantity_parameters = len(input_names)
-        fcode = data[0][0]
+        # fcode = data[0][0]
         ftype = data[0][1]
         parameter_list = []
         # exporting Python UDF Function
@@ -337,97 +255,7 @@ class Cursor(object):
             using the PDB debugger. Optionally can run on only a
             sample of the input data, for faster data export.
         """
-
-        # first gather information about the function
-        self.execute("""
-            SELECT func, type
-            FROM functions
-            WHERE language>=6 AND language <= 11 AND name='%s';""" % fname)
-        data = self.fetchall()
-        if len(data) == 0:
-            raise Exception("Function not found!")
-
-        # then gather the input arguments of the function
-        self.execute("""
-            SELECT args.name, args.type
-            FROM args
-            INNER JOIN functions ON args.func_id=functions.id
-            WHERE functions.name='%s' AND args.inout=1
-            ORDER BY args.number;""" % fname)
-        input_types = self.fetchall()
-
-        fcode = data[0][0]
-        ftype = data[0][1]
-
-        # now obtain the input columns
-        arguments = self.__exportparameters(ftype, fname, query,
-                                            len(input_types), sample)
-
-        arglist = "_columns, _column_types, _conn"
-        cleaned_arguments = dict()
-        for i in range(len(input_types)):
-            argname = "arg%d" % (i + 1)
-            if argname not in arguments:
-                raise Exception("Argument %d not found!" % (i + 1))
-            input_name = str(input_types[i][0])
-            cleaned_arguments[input_name] = arguments[argname]
-            arglist += ", %s" % input_name
-        cleaned_arguments['_columns'] = arguments['_columns']
-        cleaned_arguments['_column_types'] = arguments['_column_types']
-
-        # create a temporary file for the function execution and run it
-        with tempfile.NamedTemporaryFile() as f:
-            fcode = fcode.strip()
-            fcode = re.sub('^{', '', fcode)
-            fcode = re.sub('};$', '', fcode)
-            fcode = re.sub('^\n', '', fcode)
-            function_definition = "def pyfun(%s):\n %s\n" % (
-                arglist, fcode.replace("\n", "\n "))
-            f.write(function_definition)
-            f.flush()
-            execfile(f.name, globals(), locals())
-
-            class LoopbackObject(object):
-                def __init__(self, connection):
-                    self.__conn = connection
-
-                def execute(self, query):
-                    self.__conn.execute("""
-                        CREATE OR REPLACE FUNCTION export_parameters(*)
-                        RETURNS TABLE(s STRING) LANGUAGE PYTHON
-                        {
-                            import inspect
-                            import pickle
-                            frame = inspect.currentframe();
-                            args, _, _, values = inspect.getargvalues(frame);
-                            dd = {x: values[x] for x in args};
-                            del dd['_conn']
-                            del dd['_columns']
-                            del dd['_column_types']
-                            return pickle.dumps(dd);
-                        };""")
-                    self.__conn.execute("""
-                        SELECT *
-                        FROM (%s) AS xx
-                        LIMIT 1""" % query)
-                    query_description = self.__conn.description
-                    self.__conn.execute("""
-                        SELECT *
-                        FROM export_parameters ( (%s) );""" % query)
-                    data = self.__conn.fetchall()
-                    arguments = pickle.loads(str(data[0][0]))
-                    self.__conn.execute('DROP FUNCTION export_parameters;')
-                    if len(arguments) != len(query_description):
-                        raise Exception("Incorrect number of parameters!")
-                    result = dict()
-                    for i in range(len(arguments)):
-                        argname = "arg%d" % (i + 1)
-                        result[query_description[i][0]] = arguments[argname]
-                    return result
-
-            cleaned_arguments['_conn'] = LoopbackObject(self)
-            pdb.set_trace()
-            return locals()['pyfun'](*[], **cleaned_arguments)
+        debug(self, query, fname, sample)
 
     def fetchone(self):
         """Fetch the next row of a query result set, returning a
@@ -482,8 +310,8 @@ class Cursor(object):
         self.rownumber = min(end, len(self._rows) + self._offset)
 
         while (end > self.rownumber) and self.nextset():
-                result += self._rows[self.rownumber - self._offset:end - self._offset]
-                self.rownumber = min(end, len(self._rows) + self._offset)
+            result += self._rows[self.rownumber - self._offset:end - self._offset]
+            self.rownumber = min(end, len(self._rows) + self._offset)
         return result
 
     def fetchall(self):
@@ -615,11 +443,11 @@ class Cursor(object):
                 if identity == "name":
                     column_name = values
                 elif identity == "table_name":
-                    table_name = values   # not used
+                    _ = values   # not used
                 elif identity == "type":
                     type_ = values
                 elif identity == "length":
-                    length = values   # not used
+                    _ = values   # not used
                 elif identity == "typesizes":
                     typesizes = [[int(j) for j in i.split()] for i in values]
                     internal_size = [x[0] for x in typesizes]
