+++ setup.cfg
@@ -3,7 +3,7 @@ universal=1
 
 [tool:pytest]
 addopts= --tb native -v -r fxX --maxfail=25 -p no:warnings
-python_files=test/*test_*.py
+python_files=tests/*test_*.py
 
 [aliases]
 test=pytest
