+++ setup.py
@@ -13,6 +13,7 @@ from setuptools import setup
 def read(fname):
     return open(os.path.join(os.path.dirname(__file__), fname)).read()
 
+
 __version__ = '1.2.0'
 
 setup(name='pymonetdb',
@@ -39,10 +40,8 @@ setup(name='pymonetdb',
           "Programming Language :: Python :: 3.7",
           "Programming Language :: Python :: Implementation :: PyPy",
       ],
-      install_requires=['six'],
+      install_requires=['six', 'typing'],
       setup_requires=['pytest-runner'],
       tests_require=['pytest', 'mock'],
       test_suite="test",
-)
-
-
+      )
