+++ tests/test_capabilities.py
@@ -1,54 +1,41 @@
-#!/usr/bin/env python -O
-
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0.  If a copy of the MPL was not distributed with this
-# file, You can obtain one at http://mozilla.org/MPL/2.0/.
-#
-# Copyright 1997 - July 2008 CWI, August 2008 - 2016 MonetDB B.V.
-
 """ Script to test database capabilities and the DB-API interface
     for functionality and memory leaks.
 
     Adapted from a script by M-A Lemburg and taken from the MySQL python driver.
 
 """
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0.  If a copy of the MPL was not distributed with this
+# file, You can obtain one at http://mozilla.org/MPL/2.0/.
+#
+# Copyright 1997 - July 2008 CWI, August 2008 - 2016 MonetDB B.V.
+
+
 import datetime
-import os
 from time import time
 import unittest
-import tempfile
 
 from pymonetdb.exceptions import ProgrammingError
-import pymonetdb.sql
-from six import unichr, PY2
+import pymonetdb
+from pymonetdb.sql import monetize
+from six import unichr
 
-MAPIPORT = int(os.environ.get('MAPIPORT', 50000))
-TSTDB = os.environ.get('TSTDB', 'demo')
-TSTHOSTNAME = os.environ.get('TSTHOSTNAME', 'localhost')
-TSTUSERNAME = os.environ.get('TSTUSERNAME', 'monetdb')
-TSTPASSWORD = os.environ.get('TSTPASSWORD', 'monetdb')
+from tests.util import test_args
 
 
 class DatabaseTest(unittest.TestCase):
     create_table_extra = ''
     rows = 10
 
-    db_module = pymonetdb
-    connect_args = ()
-    connect_kwargs = dict(database=TSTDB, port=MAPIPORT, hostname=TSTHOSTNAME,
-                          username=TSTUSERNAME, password=TSTPASSWORD,
-                          autocommit=False)
     leak_test = False
 
     def setUp(self):
-        db = self.db_module.connect(*self.connect_args, **self.connect_kwargs)
+        db = pymonetdb.connect(autocommit=False, **test_args)
         self.connection = db
         self.cursor = db.cursor()
-        self.BLOBText = ''.join([chr(i) for i in range(33,127)] * 100);
-        self.BLOBBinary = self.db_module.Binary(''.join([chr(i) for i in range(256)] * 16))
+        self.BLOBText = ''.join([chr(i) for i in range(33, 127)] * 100)
+        self.BLOBBinary = pymonetdb.Binary(''.join([chr(i) for i in range(256)] * 16))
         self.BLOBUText = ''.join([unichr(i) for i in range(1, 16384)])
-        if PY2:
-            self.BLOBUText = unicode(self.BLOBUText)
 
     def tearDown(self):
         self.connection.close()
@@ -56,13 +43,14 @@ class DatabaseTest(unittest.TestCase):
     def table_exists(self, name):
         try:
             self.cursor.execute('select * from %s where 1=0' % name)
-        except:
+        except pymonetdb.OperationalError:
             self.connection.rollback()
             return False
         else:
             return True
 
-    def quote_identifier(self, ident):
+    @staticmethod
+    def quote_identifier(ident):
         return '"%s"' % ident
 
     def new_table_name(self):
@@ -88,73 +76,62 @@ class DatabaseTest(unittest.TestCase):
                              ',\n'.join(columndefs),
                              self.create_table_extra))
 
-    def check_data_integrity(self, columndefs, generator):
+    def check_data_integrity(self, columndefs, generator, drop=True):
         self.create_table(columndefs)
         insert_statement = ('INSERT INTO %s VALUES (%s)' %
                             (self.table,
                              ','.join(['%s'] * len(columndefs))))
-        data = [ [ generator(i,j) for j in range(len(columndefs)) ]
-                 for i in range(self.rows) ]
+        data = [[generator(i, j) for j in range(len(columndefs))]
+                for i in range(self.rows)]
         self.cursor.executemany(insert_statement, data)
         self.connection.commit()
-        # verify
+
         self.cursor.execute('select * from %s' % self.table)
-        l = self.cursor.fetchall()
-        self.assertEqual(len(l), self.rows)
+        rows = self.cursor.fetchall()
+        self.assertEqual(len(rows), self.rows)
         try:
             for i in range(self.rows):
                 for j in range(len(columndefs)):
-                    self.assertEqual(l[i][j], generator(i,j))
+                    self.assertEqual(rows[i][j], generator(i, j))
         finally:
-            self.cursor.execute('drop table %s' % (self.table))
+            if drop:
+                self.cursor.execute('drop table %s' % self.table)
 
     def test_transactions(self):
-        columndefs = ( 'col1 INT', 'col2 VARCHAR(255)')
-        def generator(row, col):
-            if col == 0: return row
-            else: return ('%i' % (row%10))*255
-        self.create_table(columndefs)
-        insert_statement = ('INSERT INTO %s VALUES (%s)' %
-                            (self.table,
-                             ','.join(['%s'] * len(columndefs))))
-        data = [ [ generator(i,j) for j in range(len(columndefs)) ]
-                 for i in range(self.rows) ]
-        self.cursor.executemany(insert_statement, data)
-        # verify
-        self.connection.commit()
-        self.cursor.execute('select * from %s' % self.table)
-        l = self.cursor.fetchall()
-        self.assertEqual(len(l), self.rows)
-        for i in range(self.rows):
-            for j in range(len(columndefs)):
-                self.assertEqual(l[i][j], generator(i,j))
-        delete_statement = 'delete from %s where col1=%%s' % self.table
-        self.cursor.execute(delete_statement, (0,))
-        self.cursor.execute('select col1 from %s where col1=%s' % \
-                            (self.table, 0))
-        l = self.cursor.fetchall()
-        self.assertFalse(l, "DELETE didn't work")
-        self.connection.rollback()
-        self.cursor.execute('select col1 from %s where col1=%s' % \
-                            (self.table, 0))
-        l = self.cursor.fetchall()
-        self.assertTrue(len(l) == 1, "ROLLBACK didn't work")
-        self.cursor.execute('drop table %s' % (self.table))
-
-    def test_truncation(self):
-        columndefs = ( 'col1 INT', 'col2 VARCHAR(255)')
+        columndefs = ('col1 INT', 'col2 VARCHAR(255)')
 
         def generator(row, col):
             if col == 0:
                 return row
             else:
-                return ('%i' % (row%10))*(int(255-self.rows/2)+row)
+                return ('%i' % (row % 10)) * 255
+
+        self.check_data_integrity(columndefs, generator, drop=False)
+        delete_statement = 'delete from %s where col1=%%s' % self.table
+        self.cursor.execute(delete_statement, (0,))
+        self.cursor.execute('select col1 from %s where col1=%s' % (self.table, 0))
+        rows = self.cursor.fetchall()
+        self.assertFalse(rows, "DELETE didn't work")
+        self.connection.rollback()
+        self.cursor.execute('select col1 from %s where col1=%s' % (self.table, 0))
+        rows = self.cursor.fetchall()
+        self.assertTrue(len(rows) == 1, "ROLLBACK didn't work")
+        self.cursor.execute('drop table %s' % self.table)
+
+    def test_truncation(self):
+        columndefs = ('col1 INT', 'col2 VARCHAR(255)')
+
+        def generator(row, col):
+            if col == 0:
+                return row
+            else:
+                return ('%i' % (row % 10)) * (int(255 - self.rows / 2) + row)
+
         self.create_table(columndefs)
         insert_statement = ('INSERT INTO %s VALUES (%s)' %
-                            (self.table,
-                             ','.join(['%s'] * len(columndefs))))
+                            (self.table, ','.join(['%s'] * len(columndefs))))
         try:
-            self.cursor.execute(insert_statement, (0, '0'*256))
+            self.cursor.execute(insert_statement, (0, '0' * 256))
         except Warning:
             pass
         except self.connection.Error:
@@ -169,8 +146,8 @@ class DatabaseTest(unittest.TestCase):
             for i in range(self.rows):
                 data = []
                 for j in range(len(columndefs)):
-                    data.append(generator(i,j))
-                self.cursor.execute(insert_statement,tuple(data))
+                    data.append(generator(i, j))
+                self.cursor.execute(insert_statement, tuple(data))
         except Warning:
             pass
         except self.connection.Error:
@@ -182,8 +159,8 @@ class DatabaseTest(unittest.TestCase):
         self.create_table(columndefs)
 
         try:
-            data = [ [ generator(i,j) for j in range(len(columndefs)) ]
-                     for i in range(self.rows) ]
+            data = [[generator(i, j) for j in range(len(columndefs))]
+                    for i in range(self.rows)]
             self.cursor.executemany(insert_statement, data)
         except Warning:
             pass
@@ -196,158 +173,185 @@ class DatabaseTest(unittest.TestCase):
 
     def test_CHAR(self):
         # Character data
-        def generator(row,col):
-            return ('%i' % ((row+col) % 10)) * 255
+        def generator(row, col):
+            return ('%i' % ((row + col) % 10)) * 255
+
         self.check_data_integrity(
-            ('col1 char(255)','col2 char(255)'),
+            ('col1 char(255)', 'col2 char(255)'),
             generator)
 
     def test_INT(self):
         # Number data
-        def generator(row,col):
-            return row*row
+        def generator(row, _):
+            return row * row
+
         self.check_data_integrity(
             ('col1 INT',),
             generator)
 
     def test_DECIMAL(self):
-        # DECIMAL
-        def generator(row,col):
+        def generator(row, col):
             from decimal import Decimal
             return Decimal("%d.%02d" % (row, col))
+
         self.check_data_integrity(
             ('col1 DECIMAL(5,2)',),
             generator)
 
     def test_REAL(self):
-        def generator(row,col):
-            return row*1000.0
+        def generator(row, _):
+            return row * 1000.0
+
         self.check_data_integrity(
             ('col1 REAL',),
             generator)
 
     def test_DOUBLE(self):
-        def generator(row,col):
-            return row/1e-99
+        def generator(row, _):
+            return row / 1e-99
+
         self.check_data_integrity(
             ('col1 DOUBLE',),
             generator)
 
     def test_DATE(self):
         ticks = time()
-        def generator(row,col):
-            return self.db_module.DateFromTicks(ticks+row*86400-col*1313)
+
+        def generator(row, col):
+            return pymonetdb.DateFromTicks(ticks + row * 86400 - col * 1313)
+
         self.check_data_integrity(
-                 ('col1 DATE',),
-                 generator)
+            ('col1 DATE',),
+            generator)
 
     def test_TIME(self):
         ticks = time()
-        def generator(row,col):
-            return self.db_module.TimeFromTicks(ticks+row*86400-col*1313)
+
+        def generator(row, col):
+            return pymonetdb.TimeFromTicks(ticks + row * 86400 - col * 1313)
+
         self.check_data_integrity(
-                 ('col1 TIME',),
-                 generator)
+            ('col1 TIME',),
+            generator)
 
     def test_TIMETZ(self):
         ticks = time()
-        def generator(row,col):
-            return self.db_module.TimeFromTicks(ticks+row*86400-col*1313)
+
+        def generator(row, col):
+            return pymonetdb.TimeFromTicks(ticks + row * 86400 - col * 1313)
+
         self.check_data_integrity(
-                 ('col1 TIMETZ',),
-                 generator)
+            ('col1 TIMETZ',),
+            generator)
 
     def test_DATETIME(self):
         ticks = time()
-        def generator(row,col):
-            return self.db_module.TimestampFromTicks(ticks+row*86400-col*1313)
+
+        def generator(row, col):
+            return pymonetdb.TimestampFromTicks(ticks + row * 86400 - col * 1313)
+
         self.check_data_integrity(
-                 ('col1 TIMESTAMP',),
-                 generator)
+            ('col1 TIMESTAMP',),
+            generator)
 
     def test_TIMESTAMP(self):
         ticks = time()
-        def generator(row,col):
-            return self.db_module.TimestampFromTicks(ticks+row*86400-col*1313)
+
+        def generator(row, col):
+            return pymonetdb.TimestampFromTicks(ticks + row * 86400 - col * 1313)
+
         self.check_data_integrity(
-                 ('col1 TIMESTAMP',),
-                 generator)
+            ('col1 TIMESTAMP',),
+            generator)
 
     def test_TIMESTAMPTZ(self):
         ticks = time()
-        def generator(row,col):
-            return self.db_module.TimestampFromTicks(ticks+row*86400-col*1313)
+
+        def generator(row, col):
+            return pymonetdb.TimestampFromTicks(ticks + row * 86400 - col * 1313)
+
         self.check_data_integrity(
-                 ('col1 TIMESTAMPTZ',),
-                  generator)
+            ('col1 TIMESTAMPTZ',),
+            generator)
 
     def test_fractional_TIMESTAMP(self):
         ticks = time()
-        def generator(row,col):
-            return self.db_module.TimestampFromTicks(ticks+row*86400-col*1313+row*0.7*col/3.0)
+
+        def generator(row, col):
+            return pymonetdb.TimestampFromTicks(ticks + row * 86400 - col * 1313 + row * 0.7 * col / 3.0)
+
         self.check_data_integrity(
-                 ('col1 TIMESTAMP',),
-                 generator)
+            ('col1 TIMESTAMP',),
+            generator)
 
     def test_TEXT(self):
-        def generator(row,col):
-            return self.BLOBText # 'BLOB Text ' * 1024
+        def generator(_, __):
+            return self.BLOBText  # 'BLOB Text ' * 1024
+
         self.check_data_integrity(
-                 ('col2 TEXT',),
-                 generator)
+            ('col2 TEXT',),
+            generator)
 
     def test_BLOB(self):
-        def generator(row,col):
+        def generator(row, col):
             if col == 0:
                 return row
             else:
-                return self.BLOBBinary # 'BLOB\000Binary ' * 1024
+                return self.BLOBBinary  # 'BLOB\000Binary ' * 1024
+
         self.check_data_integrity(
-                 ('col1 INT','col2 BLOB'),
-                 generator)
+            ('col1 INT', 'col2 BLOB'),
+            generator)
 
     def test_TINYINT(self):
         # Number data
-        def generator(row,col):
-            v = (row*row) % 256
+        def generator(row, _):
+            v = (row * row) % 256
             if v > 127:
-                v = v-256
+                v = v - 256
             return v
+
         self.check_data_integrity(
             ('col1 TINYINT',),
             generator)
 
     def test_small_CHAR(self):
         # Character data
-        def generator(row,col):
-            i = (row*col+62)%256
-            if i == 62: return ''
-            if i == 63: return None
+        def generator(row, col):
+            i = (row * col + 62) % 256
+            if i == 62:
+                return ''
+            if i == 63:
+                return None
             return chr(i)
+
         self.check_data_integrity(
-            ('col1 char(1)','col2 char(1)'),
+            ('col1 char(1)', 'col2 char(1)'),
             generator)
 
     def test_BOOL(self):
-        def generator(row,col):
-            return bool(row%2)
+        def generator(row, _):
+            return bool(row % 2)
+
         self.check_data_integrity(
             ('col1 BOOL',),
             generator)
 
     def test_URL(self):
-        def generator(row,col):
+        def generator(_, __):
             return "http://example.org/something"
+
         self.check_data_integrity(
-                 ('col1 URL',),
-                 generator)
+            ('col1 URL',),
+            generator)
 
     def test_INET(self):
-        def generator(row,col):
+        def generator(_, __):
             return "192.168.254.101"
+
         self.check_data_integrity(
-                 ('col1 INET',),
-                 generator)
+            ('col1 INET',),
+            generator)
 
     def test_description(self):
         self.table = self.new_table_name()
@@ -357,12 +361,13 @@ class DatabaseTest(unittest.TestCase):
             ('n', 'varchar', None, 1, None, None, None),
         ]
         try:
-            self.cursor.execute("create table %s (c VARCHAR(1024), d DECIMAL(9,4), n VARCHAR(1) NOT NULL)" % self.table);
+            self.cursor.execute(
+                "create table %s (c VARCHAR(1024), d DECIMAL(9,4), n VARCHAR(1) NOT NULL)" % self.table)
             self.cursor.execute("insert into %s VALUES ('test', 12345.1234, 'x')" % self.table)
             self.cursor.execute('select * from %s' % self.table)
             self.assertEqual(self.cursor.description, shouldbe, "cursor.description is incorrect")
         finally:
-            self.cursor.execute('drop table %s' % (self.table))
+            self.cursor.execute('drop table %s' % self.table)
 
     def test_bigresult(self):
         self.cursor.execute('select count(*) from types')
@@ -371,7 +376,7 @@ class DatabaseTest(unittest.TestCase):
         self.cursor.arraysize = 100000
         self.cursor.execute('select * from types, types')
         r = self.cursor.fetchall()
-        self.assertEqual(len(r), n**2)
+        self.assertEqual(len(r), n ** 2)
 
     def test_closecur(self):
         self.cursor.close()
@@ -380,9 +385,9 @@ class DatabaseTest(unittest.TestCase):
 
     def test_customtype(self):
         t = ["list", "test"]
-        self.assertRaises(ProgrammingError, self.db_module.sql.monetize.convert, t)
-        self.db_module.sql.monetize.mapping_dict[list] = str
-        self.assertEqual(self.db_module.sql.monetize.convert(t), "['list', 'test']")
+        self.assertRaises(ProgrammingError, monetize.convert, t)
+        monetize.mapping_dict[list] = str
+        self.assertEqual(monetize.convert(t), "['list', 'test']")
 
     def test_multiple_queries(self):
         table1 = self.new_table_name()
@@ -395,31 +400,6 @@ class DatabaseTest(unittest.TestCase):
                             (table1, table2))
         result = self.cursor.fetchall()
         self.assertEqual(result, [(50, 50)])
-
-    def test_debug_udf(self):
-        self.cursor.execute("""
-            CREATE FUNCTION test_python_udf(i INTEGER)
-            RETURNS INTEGER
-            LANGUAGE PYTHON {
-                return i * 2;
-            };
-            """)
-        # test if Python UDFs are enabled on the server
-        try:
-            self.cursor.execute('SELECT test_python_udf(1);')
-        except:
-            # python UDFs are disabled or not compiled in, skip this test
-            return
-        result = self.cursor.fetchall()
-        self.assertEqual(result, [(2,)])
-        # test python debugging capabilities
-        with tempfile.NamedTemporaryFile(delete=False) as f:
-            fname = f.name
-        self.cursor.export('SELECT test_python_udf(1)', 'test_python_udf', filespath=fname)
-        fname += "test_python_udf.py"
-        with open(fname) as f:
-            code = f.read()
-            self.assertEqual('test_python_udf(i)' in code, True)
 
     def test_temporal_operations(self):
         dt = datetime.datetime(2017, 12, 6, 12, 30)
