+++ tests/test_control.py
@@ -4,21 +4,13 @@
 #
 # Copyright 1997 - July 2008 CWI, August 2008 - 2016 MonetDB B.V.
 
-import os
 import unittest
-#import logging
-#logging.basicConfig(level=logging.DEBUG)
 from pymonetdb.control import Control
 from pymonetdb.exceptions import OperationalError
-
-FULL_TEST = False
-MAPIPORT = int(os.environ.get('MAPIPORT', 50000))
-TSTHOSTNAME = os.environ.get('TSTHOSTNAME', 'localhost')
-TSTPASSPHRASE = os.environ.get('TSTPASSPHRASE', 'testdb')
+from tests.util import test_hostname, test_port, test_passphrase, test_full
 
 database_prefix = 'controltest_'
 database_name = database_prefix + 'other'
-passphrase = 'testdb'
 
 
 def do_without_fail(function):
@@ -31,14 +23,24 @@ def do_without_fail(function):
 class TestControl(unittest.TestCase):
     def setUp(self):
         # use tcp
-        self.control = Control(TSTHOSTNAME, MAPIPORT, TSTPASSPHRASE)
+        self.control = Control(test_hostname, test_port, test_passphrase)
 
         # use socket
-        #self.control = Control()
+        # self.control = Control()
 
         do_without_fail(lambda: self.control.stop(database_name))
         do_without_fail(lambda: self.control.destroy(database_name))
         self.control.create(database_name)
+
+    def _cleanup(self, old, new):
+        do_without_fail(lambda: self.control.destroy(old))
+        do_without_fail(lambda: self.control.destroy(new))
+        self.control.create(old)
+        self.control.rename(old, new)
+        statuses = self.control.status()
+        self.assertTrue(new in [status["name"] for status in statuses])
+        do_without_fail(lambda: self.control.destroy(new))
+        return statuses
 
     def tear_down(self):
         do_without_fail(lambda: self.control.stop(database_name))
@@ -70,21 +72,16 @@ class TestControl(unittest.TestCase):
         self.assertTrue(self.control.release(database_name))
         self.assertRaises(OperationalError, self.control.release, database_name)
 
-    @unittest.skipUnless(FULL_TEST, "full test disabled")
+    @unittest.skipUnless(test_full, "full test disabled")
     def test_status(self):
         status = self.control.status(database_name)
         self.assertEqual(status["name"], database_name)
 
-    @unittest.skipUnless(FULL_TEST, "full test disabled")
+    @unittest.skipUnless(test_full, "full test disabled")
     def test_statuses(self):
         status1 = database_prefix + "status1"
         status2 = database_prefix + "status2"
-        do_without_fail(lambda: self.control.destroy(status1))
-        do_without_fail(lambda: self.control.destroy(status2))
-        self.control.create(status1)
-        self.control.create(status2)
-        statuses = self.control.status()
-        self.assertTrue(status1 in [status["name"] for status in statuses])
+        statuses = self._cleanup(status1, status2)
         self.assertTrue(status2 in [status["name"] for status in statuses])
         do_without_fail(lambda: self.control.destroy(status1))
         do_without_fail(lambda: self.control.destroy(status2))
@@ -93,12 +90,12 @@ class TestControl(unittest.TestCase):
         do_without_fail(lambda: self.control.stop(database_name))
         self.assertTrue(self.control.start(database_name))
 
-    @unittest.skipUnless(FULL_TEST, "full test disabled")
+    @unittest.skipUnless(test_full, "full test disabled")
     def test_stop(self):
         do_without_fail(lambda: self.control.start(database_name))
         self.assertTrue(self.control.stop(database_name))
 
-    @unittest.skipUnless(FULL_TEST, "full test disabled")
+    @unittest.skipUnless(test_full, "full test disabled")
     def test_kill(self):
         do_without_fail(lambda: self.control.start(database_name))
         self.assertTrue(self.control.kill(database_name))
@@ -118,26 +115,18 @@ class TestControl(unittest.TestCase):
         self.assertTrue("readonly" in self.control.get(database_name))
         self.control.inherit(database_name, "readonly")
         # TODO: False on OSX, True on travis?
-        #self.assertTrue("readonly" in self.control.get(database_name))
+        # self.assertTrue("readonly" in self.control.get(database_name))
 
     def test_rename(self):
         old = database_prefix + "old"
         new = database_prefix + "new"
-        do_without_fail(lambda: self.control.destroy(old))
-        do_without_fail(lambda: self.control.destroy(new))
-        self.control.create(old)
-        self.control.rename(old, new)
-        statuses = self.control.status()
-        self.assertTrue(new in [status["name"] for status in statuses])
+        self._cleanup(old, new)
         do_without_fail(lambda: self.control.destroy(new))
 
     def test_defaults(self):
         defaults = self.control.defaults()
         self.assertTrue("readonly" in defaults)
 
-    @unittest.skipUnless(FULL_TEST, "full test disabled")
+    @unittest.skipUnless(test_full, "full test disabled")
     def test_neighbours(self):
         self.control.neighbours()
-
-if __name__ == '__main__':
-    unittest.main()
